Changed the x_2d and y_2d members to be doubles and refactored all calculations to not lose precission with integer types, now the projection looks good. Changed the function to draw a pixel to round the doubles before drawing the pixel on the image. Added an offset in the initial scaling to make the projection slightly smaller so that doesn't get out of the screen. Added new maps to test the application. Implemented functions to draw rows and columns of the matrix.

+++ include/fdf_structures.h
@@ -29,8 +29,8 @@
  */
 typedef struct s_pixel
 {
-	int				x_2d;
-	int				y_2d;
+	double			x_2d;
+	double			y_2d;
 	int				x;
 	int				y;
 	int				z;
